move the GraphQL related code + deserialize JSON response to rust types
The hard coded GraphQL query was moved to the `graphql.rs` module, and
the newly added rust types were also added to the `graphql.rs`. In order
to deserialize into rust types we bring in `serde` and the `time` crates
as dependencies.

The `main` entrypoint was modified to use the new deserialized JSON
response.