Add `tokio` and `sqlx` to the project.
`sqlx` will allow us to issues queries to the database. Since `sqlx` is
an  async library we need to pull in a async runtime like `tokio`.

**Note**: API calls to GitHub are syncronous right now, and from some
basic testing that doesn't play well with the new async code. I didn't
want to the whole app async yet so I'll plan to address that later on.